Remove duplication of code after compiling
Previously, we made several superflous imports in the scss-files - which
results in duplicated code in the final css files (the files that is
linked in the head of each page).

Now _variables.scss is only imported in/by header.scss - since the
header will be imported in every site page.

Realising just now, maybe it would be a bit more effecient to abstract the head element to an independent blade-file, isolated from the header-blade.